+++ Tests/NablaSchedulingTests/SnapshotTests/LocationPicker/LocationPickerViewControllerTests.swift
@@ -49,6 +49,6 @@ import XCTest
         viewModel.given(.items(getter: []))
         // WHEN
         // THEN
-        assertSnapshots(matching: navigationController, as: .lightAndDarkImages())
+        assertSnapshots(matching: navigationController, as: .lightAndDarkImages(wait: 0.5))
     }
 }