+++ Sources/NablaMessagingCore/Utils/GateKeepers.swift
@@ -1,7 +1,27 @@
 import Foundation
+import NablaCore
 
-struct GateKeepers {
+// sourcery: AutoMockable
+protocol GateKeepers {
+    var supportVideoCallActionRequests: Bool { get }
+}
+
+class GateKeepersImpl: GateKeepers {
     // MARK: - Internal
     
-    let supportVideoCallActionRequests: Bool
+    var supportVideoCallActionRequests: Bool {
+        coreContainer.videoCallClient != nil
+    }
+    
+    // MARK: Init
+    
+    init(
+        coreContainer: CoreContainer
+    ) {
+        self.coreContainer = coreContainer
+    }
+
+    // MARK: - Private
+    
+    private let coreContainer: CoreContainer
 }